+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.0.1
+=====
+
+*   (bug) Properly implement `remove()` and `warmup()` for config cache items.
+*   (bug) Work around symfony's inconsistent resource tracking.
+
+
 1.0.0
 =====